refactor: split bot/i18n modules + fix ai_moderator type safety and prompt injection
Module splits:
- Extract _bot_commands callback wiring → _command_callbacks.py, _command_registration.py
- Extract _bot_plugins method scanner → _plugin_scanner.py
- Extract i18n internals → _i18n_locale.py, _i18n_diagnostics.py, _i18n_validation.py

Security (closes #48):
- Wrap message.content and message.author.name in <message>/<author> XML delimiters
  in AIModeratorPlugin._analyze_message to prevent prompt injection (audit M1 finding)

Type safety:
- Add `if ctx.guild is None: return False` guard to _execute_action (4 Pyright errors)
- Add `assert ctx.guild is not None` to all six guild_only slash handlers (8 Pyright errors)

Docs:
- Expand CLAUDE.md with architecture quick-reference, testing patterns, and key invariants
